requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
alembic
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[argon2]
pydantic
python-dotenv
pytest
httpx
pytest-asyncio